Consolidate TopBar component into separate components, create new NavMenu and ProfileMenu component, tests, and stories
