+++ src/components/Layouts/Primary/Primary.stories.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react';
 import { MockedProvider } from '@apollo/client/testing';
 import { Container, Box } from '@material-ui/core';
 import withDispatch from '../../../decorators/withDispatch';
-import { getNotificationsMocks } from './TopBar/NotificationMenu/NotificationMenu.mock';
+import { getNotificationsMocks } from './TopBar/Items/NotificationMenu/NotificationMenu.mock';
 import { getTopBarMock } from './TopBar/TopBar.mock';
 import Primary from '.';
