+++ src/components/Layouts/Primary/TopBar/Items/AddMenu/AddMenu.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import TestWrapper from '../../../../../../__tests__/util/TestWrapper';
+import TestWrapper from '../../../../../../../__tests__/util/TestWrapper';
 import AddMenu from './AddMenu';
 
 describe('AddMenu', () => {
