+++ src/components/Layouts/Primary/TopBar/Items/NotificationMenu/AcknowledgeAllUserNotifications.graphql
