+++ src/components/Layouts/Primary/TopBar/Items/NotificationMenu/NotificationMenu.stories.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import { MockedProvider } from '@apollo/client/testing';
 import { AppBar, Box } from '@material-ui/core';
-import withDispatch from '../../../../../decorators/withDispatch';
+import withDispatch from '../../../../../../decorators/withDispatch';
 import {
   getNotificationsMocks,
   getNotificationsEmptyMock,
@@ -11,7 +11,7 @@ import {
 import NotificationMenu from './NotificationMenu';
 
 export default {
-  title: 'Layouts/Primary/TopBar/NotificationMenu',
+  title: 'Layouts/Primary/TopBar/Items/NotificationMenu',
   decorators: [
     withDispatch(
       { type: 'updateAccountListId', accountListId: '1' },
