+++ src/components/Layouts/Primary/TopBar/Items/NotificationMenu/NotificationMenu.test.tsx
@@ -4,8 +4,8 @@ import { InMemoryCache } from '@apollo/client';
 import {
   render,
   waitFor,
-} from '../../../../../../__tests__/util/testingLibraryReactMock';
-import TestWrapper from '../../../../../../__tests__/util/TestWrapper';
+} from '../../../../../../../__tests__/util/testingLibraryReactMock';
+import TestWrapper from '../../../../../../../__tests__/util/TestWrapper';
 import {
   acknowledgeAllUserNotificationsMutationMock,
   getNotificationsEmptyMock,
