+++ src/components/Layouts/Primary/TopBar/Items/NotificationMenu/NotificationMenu.tsx
@@ -13,8 +13,8 @@ import React, { ReactElement, useEffect, useState } from 'react';
 import NotificationsNoneIcon from '@material-ui/icons/NotificationsNone';
 import NotificationsIcon from '@material-ui/icons/Notifications';
 import { useTranslation } from 'react-i18next';
-import { useApp } from '../../../../App';
-import illustration13 from '../../../../../images/drawkit/grape/drawkit-grape-pack-illustration-13.svg';
+import { useApp } from '../../../../../App';
+import illustration13 from '../../../../../../images/drawkit/grape/drawkit-grape-pack-illustration-13.svg';
 import NotificationMenuItem from './Item';
 import {
   GetNotificationsDocument,
