+++ src/components/Layouts/Primary/TopBar/Items/SearchMenu/SearchMenu.stories.tsx
@@ -1,10 +1,10 @@
 import React, { ReactElement } from 'react';
 import { AppBar, Box } from '@material-ui/core';
-import withDispatch from '../../../../../decorators/withDispatch';
-import AddMenu from './AddMenu';
+import withDispatch from '../../../../../../decorators/withDispatch';
+import SearchMenu from './SearchMenu';
 
 export default {
-  title: 'Layouts/Primary/TopBar/AddMenu',
+  title: 'Layouts/Primary/TopBar/Items/SearchMenu',
   decorators: [
     withDispatch({ type: 'updateBreadcrumb', breadcrumb: 'Dashboard' }),
   ],
@@ -15,7 +15,7 @@ export const Default = (): ReactElement => {
     <>
       <AppBar>
         <Box>
-          <AddMenu />
+          <SearchMenu />
         </Box>
       </AppBar>
     </>
