+++ src/components/Layouts/Primary/TopBar/Items/SearchMenu/SearchMenu.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import TestWrapper from '../../../../../../__tests__/util/TestWrapper';
+import TestWrapper from '../../../../../../../__tests__/util/TestWrapper';
 import SearchMenu from './SearchMenu';
 
 describe('SearchMenu', () => {
