+++ src/components/Layouts/Primary/TopBar/TopBar.tsx
@@ -1,11 +1,6 @@
-import React, { ReactElement, useState, useEffect } from 'react';
+import React, { ReactElement } from 'react';
 import {
-  Avatar,
-  IconButton,
-  Button,
-  MenuItem,
   Box,
-  Menu,
   makeStyles,
   Toolbar,
   AppBar,
@@ -13,23 +8,14 @@ import {
   Theme,
   Grid,
   Hidden,
-  ListItemText,
-  Divider,
-  ListItemAvatar,
-  Link,
 } from '@material-ui/core';
-import { useTranslation } from 'react-i18next';
-import { signout } from 'next-auth/client';
-import NextLink from 'next/link';
-import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import { useApp } from '../../../App';
-import HandoffLink from '../../../HandoffLink';
-import { User } from '../../../../../graphql/types.generated';
 import logo from '../../../../images/logo.svg';
-import NotificationMenu from './NotificationMenu/NotificationMenu';
-import AddMenu from './AddMenu/AddMenu';
-import { useGetTopBarQuery } from './GetTopBar.generated';
-import SearchMenu from './SearchMenu/SearchMenu';
+import { useApp } from '../../../App';
+import NotificationMenu from './Items/NotificationMenu/NotificationMenu';
+import AddMenu from './Items/AddMenu/AddMenu';
+import SearchMenu from './Items/SearchMenu/SearchMenu';
+import NavMenu from './Items/NavMenu/NavMenu';
+import ProfileMenu from './Items/ProfileMenu/ProfileMenu';
 
 const useStyles = makeStyles((theme: Theme) => ({
   appBar: {
@@ -45,12 +31,6 @@ const useStyles = makeStyles((theme: Theme) => ({
   logoGrid: {
     order: 1,
   },
-  navListItem: {
-    order: 2,
-    [theme.breakpoints.down('sm')]: {
-      display: 'none',
-    },
-  },
   addMenuGrid: {
     order: 5,
     [theme.breakpoints.down('xs')]: {
@@ -63,42 +43,11 @@ const useStyles = makeStyles((theme: Theme) => ({
       display: 'none',
     },
   },
-  notificationsGrid: {
+  notificationMenuGrid: {
     order: 6,
   },
-  avatarGrid: {
+  profileMenuGrid: {
     order: 7,
-  },
-  avatar: {
-    color: theme.palette.secondary.dark,
-  },
-  link: {
-    textTransform: 'none',
-    color: 'rgba(255,255,255,0.75)',
-    transition: 'color 0.2s ease-in-out',
-    '&:hover': {
-      color: 'rgba(255,255,255,1)',
-    },
-  },
-  button: {
-    textTransform: 'none',
-  },
-  menuList: {
-    paddingTop: 0,
-  },
-  menuItemAccount: {
-    paddingTop: 0,
-    outline: 0,
-  },
-  menuItemFooter: {
-    fontSize: theme.typography.body2.fontSize,
-    justifyContent: 'center',
-    paddingTop: theme.spacing(2),
-    outline: 0,
-  },
-  menuButton: {
-    width: '100%',
-    marginTop: theme.spacing(1),
   },
   logo: {
     width: 90,
@@ -107,45 +56,22 @@ const useStyles = makeStyles((theme: Theme) => ({
     }),
     marginRight: theme.spacing(4),
   },
-  accountName: {
-    color: '#FFFFFF',
-    padding: '0px 8px',
-  },
 }));
 
 const TopBar = (): ReactElement => {
   const classes = useStyles();
-  const { dispatch, state } = useApp();
-  const { t } = useTranslation();
+  const { state } = useApp();
   const trigger = useScrollTrigger({
     disableHysteresis: true,
     threshold: 0,
   });
-  const { data } = useGetTopBarQuery();
-  const [profileMenuAnchorEl, setProfileMenuAnchorEl] = useState(null);
-  const profileMenuOpen = Boolean(profileMenuAnchorEl);
-
-  const handleProfileMenuOpen = (event) => {
-    setProfileMenuAnchorEl(event.currentTarget);
-  };
-
-  const handleProfileMenuClose = () => {
-    setProfileMenuAnchorEl(null);
-  };
-
-  useEffect(() => {
-    data?.user &&
-      state.user?.id !== data.user.id &&
-      // TODO: Use fragments to ensure all required fields are loaded
-      dispatch({ type: 'updateUser', user: data.user as User });
-  }, [data?.user]);
 
   return (
     <>
       <AppBar className={classes.appBar} elevation={trigger ? 3 : 0}>
         <Toolbar className={classes.toolbar}>
           <Grid container alignItems="center">
-            <Grid container alignItems="center" xs="auto" md={7}>
+            <Grid container alignItems="center" xs="auto" md={1}>
               <Grid item className={classes.logoGrid}>
                 <Hidden smDown>
                   <Box className={classes.logo}>
@@ -153,57 +79,11 @@ const TopBar = (): ReactElement => {
                   </Box>
                 </Hidden>
               </Grid>
-              {state.accountListId ? (
-                <>
-                  <Grid item className={classes.navListItem}>
-                    <NextLink
-                      href="/accountLists/[accountListId]"
-                      as={`/accountLists/${state.accountListId}`}
-                      scroll={false}
-                    >
-                      <MenuItem>
-                        <ListItemText primary={t('Dashboard')} />
-                      </MenuItem>
-                    </NextLink>
-                  </Grid>
-                  <Grid item className={classes.navListItem}>
-                    <NextLink
-                      href="/accountLists/[accountListId]/contacts"
-                      as={`/accountLists/${state.accountListId}/contacts`}
-                      scroll={false}
-                    >
-                      <MenuItem>
-                        <ListItemText primary={t('Contacts')} />
-                      </MenuItem>
-                    </NextLink>
-                  </Grid>
-                  <Grid item className={classes.navListItem}>
-                    <HandoffLink path="/reports">
-                      <MenuItem onClick={handleProfileMenuClose} component="a">
-                        <ListItemText primary={t('Reports')} />
-                      </MenuItem>
-                    </HandoffLink>
-                  </Grid>
-                  <Grid item className={classes.navListItem}>
-                    <HandoffLink path="/tools">
-                      <MenuItem onClick={handleProfileMenuClose} component="a">
-                        <ListItemText primary={t('Tools')} />
-                      </MenuItem>
-                    </HandoffLink>
-                  </Grid>
-                  <Grid item className={classes.navListItem}>
-                    <HandoffLink path="/coaches">
-                      <MenuItem onClick={handleProfileMenuClose} component="a">
-                        <ListItemText primary={t('Coaches')} />
-                      </MenuItem>
-                    </HandoffLink>
-                  </Grid>
-                </>
-              ) : null}
             </Grid>
+            <NavMenu />
             <Grid
               xs={12}
-              md={5}
+              md={state.accountListId ? 5 : 11}
               container
               alignItems="center"
               justify="flex-end"
@@ -214,131 +94,16 @@ const TopBar = (): ReactElement => {
               <Grid item className={classes.addMenuGrid}>
                 <AddMenu />
               </Grid>
-              <Grid item className={classes.notificationsGrid}>
+              <Grid item className={classes.notificationMenuGrid}>
                 <NotificationMenu />
               </Grid>
-              <Grid item className={classes.avatarGrid}>
-                <IconButton
-                  onClick={handleProfileMenuOpen}
-                  data-testid="profileMenuButton"
-                >
-                  <AccountCircleIcon className={classes.avatar} />
-                  {data && (
-                    <ListItemText
-                      className={classes.accountName}
-                      primary={[data.user.firstName, data.user.lastName]
-                        .filter(Boolean)
-                        .join(' ')}
-                    />
-                  )}
-                </IconButton>
+              <Grid item className={classes.profileMenuGrid}>
+                <ProfileMenu />
               </Grid>
             </Grid>
           </Grid>
         </Toolbar>
       </AppBar>
-      <Menu
-        data-testid="profileMenu"
-        anchorEl={profileMenuAnchorEl}
-        open={profileMenuOpen}
-        onClose={handleProfileMenuClose}
-      >
-        {data && (
-          <MenuItem button={false} className={classes.menuItemAccount}>
-            <ListItemAvatar>
-              <Avatar>{data.user.firstName[0]}</Avatar>
-            </ListItemAvatar>
-            <ListItemText
-              primary={[data.user.firstName, data.user.lastName]
-                .filter(Boolean)
-                .join(' ')}
-              secondary={data.user.keyAccounts[0]?.email}
-            />
-          </MenuItem>
-        )}
-        <Divider />
-        <HandoffLink path="/preferences/personal">
-          <MenuItem onClick={handleProfileMenuClose} component="a">
-            <ListItemText primary={t('Preferences')} />
-          </MenuItem>
-        </HandoffLink>
-        <HandoffLink path="/preferences/notifications">
-          <MenuItem onClick={handleProfileMenuClose} component="a">
-            <ListItemText primary={t('Notifications')} />
-          </MenuItem>
-        </HandoffLink>
-        <HandoffLink path="/preferences/integrations">
-          <MenuItem onClick={handleProfileMenuClose} component="a">
-            <ListItemText primary={t('Connect Services')} />
-          </MenuItem>
-        </HandoffLink>
-        <HandoffLink path="/preferences/accounts">
-          <MenuItem onClick={handleProfileMenuClose} component="a">
-            <ListItemText primary={t('Manage Accounts')} />
-          </MenuItem>
-        </HandoffLink>
-        <HandoffLink path="/preferences/coaches">
-          <MenuItem onClick={handleProfileMenuClose} component="a">
-            <ListItemText primary={t('Manage Coaches')} />
-          </MenuItem>
-        </HandoffLink>
-        {(data?.user?.admin ||
-          !!data?.user?.administrativeOrganizations?.nodes?.length) && (
-          <HandoffLink path="/preferences/organizations">
-            <MenuItem onClick={handleProfileMenuClose} component="a">
-              <ListItemText primary={t('Manage Organizations')} />
-            </MenuItem>
-          </HandoffLink>
-        )}
-        {(data?.user?.admin || data?.user?.developer) && (
-          <HandoffLink path="/preferences/admin">
-            <MenuItem onClick={handleProfileMenuClose} component="a">
-              <ListItemText primary={t('Admin Console')} />
-            </MenuItem>
-          </HandoffLink>
-        )}
-        {data?.user?.developer && (
-          <HandoffLink path="/auth/user/admin" auth>
-            <MenuItem onClick={handleProfileMenuClose} component="a">
-              <ListItemText primary={t('Backend Admin')} />
-            </MenuItem>
-          </HandoffLink>
-        )}
-        {data?.user?.developer && (
-          <HandoffLink path="/auth/user/sidekiq" auth>
-            <MenuItem onClick={handleProfileMenuClose} component="a">
-              <ListItemText primary={t('Sidekiq')} />
-            </MenuItem>
-          </HandoffLink>
-        )}
-        <MenuItem button={false}>
-          <Button
-            className={classes.menuButton}
-            variant="outlined"
-            color="default"
-            onClick={() => signout()}
-          >
-            {t('Sign Out')}
-          </Button>
-        </MenuItem>
-        <MenuItem button={false} className={classes.menuItemFooter}>
-          <Link
-            href="https://get.mpdx.org/privacy-policy/"
-            target="_blank"
-            onClick={handleProfileMenuClose}
-          >
-            {t('Privacy Policy')}
-          </Link>
-          &nbsp; • &nbsp;
-          <Link
-            href="https://get.mpdx.org/terms-of-use/"
-            target="_blank"
-            onClick={handleProfileMenuClose}
-          >
-            {t('Terms of Use')}
-          </Link>
-        </MenuItem>
-      </Menu>
     </>
   );
 };
